:recycle: ページ構成を変更
ページ構成を変更しました。

- /
  - 認証画面
- /dashboard
  - 認証後のホーム画面
